Filter ABI if exported include dir is specified.
The ABI dumper used to disable filtering ABI when no header file is
found in the exported include dirs. The output content depends on the
existence of the header files.

This commit makes the dumper decide the filtering by the command line
arguments. If at least one exported include dir is specifed, the dumper
filters the ABI, so that it produces consistent output in an environment
where some usused header files don't exist.

Test: development/vndk/tools/header-checker/tests/test.py
Bug: 153926507
Change-Id: I5c1fd1e74554d875554b4908c9caaf28b9e73236